[develop -> master] 크래프팅 박스 추가 등

+++ Source/BackStreet/Item/public/ItemBase.h
@@ -24,7 +24,7 @@ public:
 	FDeleSpawnMissionItem Dele_MissionItemSpawned;
 
 	UFUNCTION(BlueprintCallable)
-		static int32 GetTargetItemKey(int32 Type, uint8 ItemID) { return Type * 1000 + ItemID; }
+		static int32 GetTargetItemKey(int32 Type, int32 ItemID) { return Type * 1000 + ItemID; }
 
 // ------ Global, Component ---------------------------------------------
 public:
@@ -59,12 +59,12 @@ protected:
 		EItemCategoryInfo ItemType = EItemCategoryInfo::E_None;
 
 	UPROPERTY(EditDefaultsOnly, BlueprintReadOnly, Category = "Gameplay")
-		uint8 ItemID;
+		int32 ItemID;
 // ------ 아이템 기본 로직-------------------------------------
 public:
 	// 외부에서 Init하기위해 Call
 	UFUNCTION()
-		void InitItem(EItemCategoryInfo SetType, uint8 NewItemID = 0);
+		void InitItem(EItemCategoryInfo SetType, int32 NewItemID = 0);
 
 	//아이템 초기 효과를 출력하고 활성화 시킨다.
 	UFUNCTION(BlueprintImplementableEvent)
